Back-out the davenport gem for now.
I'm going to need to carve that out as a separate task.
The gem install on Travis fails because it doesn't have the C library.
Same will happen to anyone else but me.
Most likely need work on the Gem to have the native install also build
the Davenport C library if it isn't already present.
Either that or install it on Travis as part of setup.